aichat_py_root/web_search_rag/requirements.txt
requests
beautifulsoup4
langchain
langchain-community
langchain-ollama
faiss-cpu
pyyaml
